+++ cmd/forklift/main.go
@@ -12,6 +12,7 @@ import (
 	"github.com/PlanktoScope/forklift/cmd/forklift/dev"
 	"github.com/PlanktoScope/forklift/cmd/forklift/host"
 	"github.com/PlanktoScope/forklift/cmd/forklift/plt"
+	"github.com/PlanktoScope/forklift/cmd/forklift/stage"
 )
 
 func main() {
@@ -28,6 +29,7 @@ var app = &cli.App{
 	Usage:   "Manages pallets and package deployments",
 	Commands: []*cli.Command{
 		plt.MakeCmd(toolVersion, repoMinVersion, palletMinVersion, newStageVersion),
+		stage.MakeCmd(toolVersion, stageMinVersion),
 		cache.Cmd,
 		host.Cmd,
 		dev.MakeCmd(toolVersion, repoMinVersion, palletMinVersion, newStageVersion),
@@ -59,9 +61,12 @@ const (
 	// palletMinVersion is the minimum supported Forklift version among pallets. A pallet with a
 	// lower Forklift version cannot be used.
 	palletMinVersion = "v0.4.0"
-	// newStageVersion is the Forklift version reported in new staged pallets made by Forklift. Older
-	// versions of the Forklift tool cannot use such staged pallets.
-	newStageVersion = "v0.7.0"
+	// palletMinVersion is the minimum supported Forklift version among staged pallet bundles. A
+	// bundle with a lower Forklift version cannot be used.
+	stageMinVersion = "v0.7.0-dev"
+	// newStageVersion is the Forklift version reported in new staged pallet bundles made by Forklift.
+	// Older versions of the Forklift tool cannot use such bundles.
+	newStageVersion = "v0.7.0-dev"
 	// fallbackVersion is the version reported which the Forklift tool reports itself as if its actual
 	// version is unknown.
 	fallbackVersion = "v0.7.0-dev"